+++ components/menu/resources/ResourceDisplayComponent.jsx
