+++ components/menu/resources/ResourceIconComponent.jsx
